+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Campaign Changelog
 
+## Unreleased
+### Added
+- Added call to queue pending sendouts after CP login
+
 ## 1.0.0-beta7 - 2018-04-05
 ### Changed
 - Changed Craft version requirement to 3.0.1